# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'lciii-scaffolding'","description":"'LCII Initial Scaffolding'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'LCII Initial Scaffolding'"}],"paths":{"/players":{"post":{"tags":["player-controller"],"summary":"Create a new player","description":"Return the player created with your id. If a player exist with the username or email, then return 404. Additionally, the email must be valid and the password must have at least 8 characters and contain at least one number, one lowercase letter, one uppercase letter and one special character.","operationId":"savePlayer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewPlayerRequestDTO"}}},"required":true},"responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlayerResponseDTO"}}}},"400":{"description":"Username or email already exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/matches":{"post":{"tags":["match-controller"],"operationId":"saveMatch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewMatchRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MatchResponseDTO"}}}}}}},"/matches/{id}/plays/":{"post":{"tags":["match-controller"],"operationId":"saveMatch_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlayRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlayResponseDTO"}}}}}}},"/login":{"post":{"tags":["login-controller"],"summary":"Login a player in the platform","description":"Return the player logged in if the credentials are correct.","operationId":"loginPlayer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginBody"}}},"required":true},"responses":{"404":{"description":"The credentials are invalid","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/players/{id}":{"get":{"tags":["player-controller"],"summary":"Get a player by id","description":"Return a player by they id. If the player doesn't exist return 404","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlayerResponseDTO"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/players/{id}/matches":{"get":{"tags":["player-controller"],"operationId":"getMatchesOfPlayer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MatchResponseDTO"}}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/matches/{id}":{"get":{"tags":["match-controller"],"operationId":"getById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MatchResponseDTO"}}}}}}}},"components":{"schemas":{"NewPlayerRequestDTO":{"type":"object","properties":{"userName":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"}}},"PlayerResponseDTO":{"type":"object","properties":{"userName":{"type":"string"},"avatar":{"type":"string"},"balance":{"type":"number"}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}},"NewMatchRequestDTO":{"required":["playerOneId","playerTwoId"],"type":"object","properties":{"playerOneId":{"type":"integer","format":"int64"},"playerTwoId":{"type":"integer","format":"int64"}}},"Card":{"type":"object","properties":{"cardSuit":{"type":"string","enum":["OROS","COPAS","ESPADAS","BASTOS"]},"number":{"type":"integer","format":"int32"},"value":{"type":"number"}}},"MatchResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"lastCard":{"$ref":"#/components/schemas/Card"},"nextCardIndex":{"type":"integer","format":"int32"},"playerOne":{"$ref":"#/components/schemas/PlayerResponseDTO"},"playerTwo":{"$ref":"#/components/schemas/PlayerResponseDTO"},"nextToPlay":{"$ref":"#/components/schemas/PlayerResponseDTO"},"winner":{"$ref":"#/components/schemas/PlayerResponseDTO"},"matchStatus":{"type":"string","enum":["PLAYING","FINISH"]}}},"PlayRequestDTO":{"required":["decision","player"],"type":"object","properties":{"player":{"type":"integer","format":"int64"},"decision":{"type":"string","enum":["MAJOR","MINOR"]}}},"PlayResponseDTO":{"type":"object","properties":{"player":{"type":"integer","format":"int64"},"decision":{"type":"string","enum":["MAJOR","MINOR"]},"previous_card":{"$ref":"#/components/schemas/Card"},"your_card":{"$ref":"#/components/schemas/Card"},"number_of_cards_in_deck":{"type":"integer","format":"int32"},"match_status":{"type":"string","enum":["PLAYING","FINISH"]}}},"LoginBody":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"Player":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"userName":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"},"avatar":{"type":"string"},"balance":{"type":"number"}}}}}}